Slight refactoring of the factory methods
In preparation for adding new network types.
